api/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
openai>=1.40.0
faiss-cpu==1.8.0
numpy<2
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-core>=0.3.0,<0.4.0
langgraph>=0.2.27,<0.3.0
pydantic>=2.7.4,<3.0.0
yfinance==0.2.36
delta-spark==3.1.0
pyspark==3.5.1
tiktoken